To asses the performance of asphalt mixtures due to the application of dynamic loading on the flexible pavement with heterogeneous asphalt layer.
Since the asphalt layer carries the maximum amount of stress when the pavement is subjected to loading. So it is important to calculate accurate response of asphalt layer for effective design. To make the design simple, IRC considered the asphalt layer as homogenous but in reality, it is not the case. For designing of asphalt layer, it is assumed as homogenous and a value of resilient modulus is taken as input. While doing so the heterogeneous nature of the asphalt layer is neglected which makes this design insufficient. The heterogeneous nature of asphalt layer is considered by taking spatially varied values of instantaneous resilient modulus as input. This pavement is modelled in ABAQUS and the response is measured.